Factor out ApplicationLoader to load startup textures etc.
