+++ libs/s25main/RttrForeachPt.h
@@ -17,8 +17,8 @@
 
 #pragma once
 
-#ifndef MACROS_H_INCLUDED
-#define MACROS_H_INCLUDED
+#ifndef RTTR_FOREACH_PT_H_INCLUDED
+#define RTTR_FOREACH_PT_H_INCLUDED
 
 /// Iterate over all points of an area using a point of TYPE named "pt"
 /// WIDTH and HEIGHT is evaluated at most once
@@ -34,4 +34,4 @@
             for(TYPE pt(0, 0); pt.y < rttrForeachPtHeight; ++pt.y)              \
                 for(pt.x = 0; pt.x < rttrForeachPtWidth; ++pt.x)
 
-#endif // !MACROS_H_INCLUDED
+#endif // !RTTR_FOREACH_PT_H_INCLUDED
